Clean up directory structure
